Reorganize client test folders
There was a useless layer in the folder structure with `/browser/unit`
so the unit has been removed and all paths in tests have been adjusted.

It was also confusing to have the folder named `browser` but the test
command named `test:client` so the command has been renamed
`test:browser` and all corresponding commands have been updated
