+++ package.json
@@ -78,9 +78,9 @@
     "watch:client:sass": "node-sass -w -r assets/styles/*.scss -o assets/styles",
     "watch:client:autoprefix": "npm run sass && postcss --use autoprefixer -r -w assets/styles/styles.css assets/styles/styles.css",
     "watch:client:js": "watchify --debug -t brfs assets/app/app.js -o assets/js/bundle.js",
-    "test": "npm run build && npm run test:server && npm run test:client && npm run test:integration",
+    "test": "npm run build && npm run test:server && npm run test:browser && npm run test:integration",
     "test:server": "NODE_ENV=test mocha --timeout 60000 test/api/bootstrap.test.js test/api/unit/**/*.test.js",
-    "test:client": "NODE_ENV=test mocha test/browser/bootstrap.test.js \"test/browser/unit/**/*.test.js\"",
+    "test:browser": "NODE_ENV=test mocha test/browser/bootstrap.test.js \"test/browser/**/*.test.js\"",
     "test:integration": "NODE_ENV=test mocha test/integration/bootstrap.test.js \"test/integration/**/*.test.js\"",
     "test:selenium:install": "node_modules/selenium-standalone/bin/selenium-standalone install",
     "test:selenium:start": "node_modules/selenium-standalone/bin/selenium-standalone start",
