+++ test/browser/models/Github.test.js
@@ -9,7 +9,7 @@ var helpers = require('./githubHelpers');
 var mockData = JSON.stringify(require('../data/repoResponse.json'));
 var mockCommitResponse = JSON.stringify(require('../data/commitResponse.json'));
 
-var Github = require('./../../../../assets/app/models/Github');
+var Github = require('./../../../assets/app/models/Github');
 
 var server;
