+++ test/browser/views/editor/breadcrumb.test.js
@@ -5,14 +5,14 @@ var sinon = require('sinon');
 var mockData = JSON.stringify(require('../../data/repoResponse.json'));
 var mockCommitResponse = JSON.stringify(require('../../data/commitResponse.json'));
 
-var Github = require('./../../../../../assets/app/models/Github');
+var Github = require('./../../../../assets/app/models/Github');
 var githubHelpers = require('../../models/githubHelpers');
 
-var PagesView = require('./../../../../../assets/app/views/editor/pages');
+var BreadcrumbView = require('./../../../../assets/app/views/editor/breadcrumb');
 
 var server;
 
-describe('pages view', function () {
+describe('breadcrumb view', function () {
   var model;
 
   beforeEach(function () {
@@ -30,23 +30,15 @@ describe('pages view', function () {
     describe('without model', function () {
       it('should throw exception', function () {
         assert.throws(function () {
-          new PagesView();
+          new BreadcrumbView();
         });
       });
     });
 
-    describe('with a model but without a pages option', function () {
-      it('should throw exception', function () {
-        assert.throws(function () {
-          new PagesView({ model: model });
-        });
-      });
-    });
-
-    describe('with a model and a pages option', function () {
+    describe('with a model', function () {
       it('should not throw exception', function () {
         assert.doesNotThrow(function () {
-          new PagesView({ model: model, pages: [] });
+          new BreadcrumbView({ model: model });
         });
       });
     });
